feat: 근무 타입 추가 및 ts 변경, 회원가입 레이아웃 수정 (#55)
