+++ src/components/register/RegisterForm.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import * as S from '@/pages/Register/Register.styles';
+import * as S from '@/components/Register/Register.styles';
 import {
 	FormData,
 	RegisterFormErrors,
@@ -7,6 +7,7 @@ import {
 	ROLE_OPTIONS,
 	POSITION_OPTIONS,
 	GENDER_OPTIONS,
+	SHIFT_TYPE_OPTIONS,
 } from '@/types/register';
 import { validateForm } from './FormValidation';
 
@@ -19,7 +20,9 @@ export function RegisterForm({ onSubmit, isSubmitting, submitError }: RegisterFo
 		gender: '',
 		age: '',
 		position: '',
-		workingHours: '',
+		userName: '',
+		userAlias: '',
+		shiftType: '',
 	});
 	const [errors, setErrors] = useState<RegisterFormErrors>({});
 
@@ -100,25 +103,6 @@ export function RegisterForm({ onSubmit, isSubmitting, submitError }: RegisterFo
 						<S.FormField>
 							<S.Label htmlFor="role">역할</S.Label>
 							<S.Select
-								id="role"
-								name="role"
-								value={formData.role}
-								onChange={handleInputChange}
-								$validation={errors.role ? 'invalid' : 'default'}
-							>
-								<option value="">선택하세요</option>
-								{Object.values(ROLE_OPTIONS).map((role) => (
-									<option key={role.value} value={role.value}>
-										{role.label}
-									</option>
-								))}
-							</S.Select>
-							{errors.role && <S.ErrorMessage>{errors.role}</S.ErrorMessage>}
-						</S.FormField>
-
-						<S.FormField>
-							<S.Label htmlFor="position">직책</S.Label>
-							<S.Select
 								id="position"
 								name="position"
 								value={formData.position}
@@ -136,6 +120,44 @@ export function RegisterForm({ onSubmit, isSubmitting, submitError }: RegisterFo
 						</S.FormField>
 
 						<S.FormField>
+							<S.Label htmlFor="shiftType">시간 타입</S.Label>
+							<S.Select
+								id="shiftType"
+								name="shiftType"
+								value={formData.shiftType}
+								onChange={handleInputChange}
+								$validation={errors.gender ? 'invalid' : 'default'}
+							>
+								<option value="">선택하세요</option>
+								{Object.values(SHIFT_TYPE_OPTIONS).map((shiftType) => (
+									<option key={shiftType.label} value={shiftType.value}>
+										{shiftType.label}
+									</option>
+								))}
+							</S.Select>
+							{errors.gender && <S.ErrorMessage>{errors.gender}</S.ErrorMessage>}
+						</S.FormField>
+
+						<S.FormField>
+							<S.Label htmlFor="role">직책</S.Label>
+							<S.Select
+								id="role"
+								name="role"
+								value={formData.role}
+								onChange={handleInputChange}
+								$validation={errors.role ? 'invalid' : 'default'}
+							>
+								<option value="">선택하세요</option>
+								{Object.values(ROLE_OPTIONS).map((role) => (
+									<option key={role.label} value={role.value}>
+										{role.label}
+									</option>
+								))}
+							</S.Select>
+							{errors.position && <S.ErrorMessage>{errors.position}</S.ErrorMessage>}
+						</S.FormField>
+
+						<S.FormField>
 							<S.Label htmlFor="gender">성별</S.Label>
 							<S.Select
 								id="gender"
@@ -145,9 +167,9 @@ export function RegisterForm({ onSubmit, isSubmitting, submitError }: RegisterFo
 								$validation={errors.gender ? 'invalid' : 'default'}
 							>
 								<option value="">선택하세요</option>
-								{GENDER_OPTIONS.map((gender) => (
-									<option key={gender} value={gender}>
-										{gender}
+								{Object.values(GENDER_OPTIONS).map((gender) => (
+									<option key={gender.label} value={gender.value}>
+										{gender.label}
 									</option>
 								))}
 							</S.Select>
